Replace Lottie with SpriteSheet-based rendered. 5x+ CPU usage decrease
At startup (or whenever dimensions change) we render the individual
frames of the Lottie animation to a SpriteSheet, and then later play
the Sprites from the Sheet, rather than recreating the animation itself.

Note something you would normally bother with, but this animation may
run for a long time, so its CPU usage becomes directly relevant to
battery life.

Obviously this comes at the cost of a few MB of RAM for the SpriteSheet
cache.

Closes #9
Closes #14